backend/pyproject.toml
[project]
name = "extreme-temps"
version = "0.2.0"
description = "Historical weather extremes analysis — data pipeline, compute engine, and API"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.1",
    "fastapi>=0.115",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pandas>=2.0",
    "pyarrow",
    "requests",
    "numpy",
    "scipy",
    "google-cloud-bigquery",
    "db-dtypes",
]

[project.scripts]
extreme-temps = "extreme_temps.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extreme_temps"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]
